requirements.txt
streamlit
groq
tavily-python
requests
beautifulsoup4
lxml
gspread
google-auth